check critical training

+++ src/meshed-memory-transformer/models/beam_search/beam_search.py
@@ -40,7 +40,7 @@ class BeamSearch(object):
             print("other", beam.expand(*([self.b_s, self.beam_size] + shape[1:])))
 
             s = torch.gather(s.view(*([self.b_s, cur_beam_size] + shape[1:])), 1,
-                             beam.expand(*([self.b_s, self.beam_size] + shape[1:])))
+                             beam.expand(*([self.b_s, self.beam_size] + shape[1:])).long())
             s = s.view(*([-1, ] + shape[1:]))
             return s
